% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_bioactivities.R
\name{Bioactivities}
\alias{Bioactivities}
\title{Bioactivities class}
\usage{
Bioactivities(.target_id = NULL, .standard_type = NULL, .data = NULL)
}
\arguments{
\item{.target_id}{target identifiers (character vector). Default: NULL}

\item{.standard_type}{standard type identifier (character). Default: NULL}

\item{.data}{bioactivities data (data.frame). Default: NULL}
}
\value{
A \code{Bioactivities} object.
}
\description{
Creates a \code{Bioactivities} object.
}
