% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-building.R
\name{build_qsar_models}
\alias{build_qsar_models}
\title{Build QSAR models with k-fold cross-validation}
\usage{
build_qsar_models(data_file, k = 5)
}
\arguments{
\item{data_file}{The file path of the dataset.}

\item{k}{The number of folds for cross-validation (default is 5).}
}
\value{
A list containing MLR, PLS, and Random Forest models with their predictions, actuals, and formulas.
}
\description{
This function builds QSAR (Quantitative Structure-Activity Relationship) models
using multiple algorithms such as Multiple Linear Regression (MLR), Partial Least Squares (PLS),
and Random Forest with k-fold cross-validation.
}
