% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation-plot.R
\name{correlation_plots}
\alias{correlation_plots}
\title{Create correlation plots for QSAR models}
\usage{
correlation_plots(model_results)
}
\arguments{
\item{model_results}{A list containing QSAR model results.}
}
\value{
A list of correlation plots for each QSAR model.
}
\description{
This function creates correlation plots for QSAR models, showing the relationship
between predicted and actual values with a correlation coefficient.
}
