% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_datasets.R
\name{createQSARDatasets}
\alias{createQSARDatasets}
\title{Creates QSAR datasets}
\usage{
createQSARDatasets(.connection, .targets, .molrepresentation, .response,
  .nrow_min = NULL, .nrow_max = NULL)
}
\arguments{
\item{.connection}{a database connection created using \code{dplyr::src_*}.}

\item{.targets}{a \code{Targets} object.}

\item{.molrepresentation}{a \code{MolRepresentation} object.}

\item{.response}{name of the response variable (character).}

\item{.nrow_min}{min number of molecules per dataset (numeric). Default: NULL.}

\item{.nrow_max}{max number of molecules per dataset (numeric). Default: NULL.}
}
\value{
A \code{QSARDatasets} object.
}
\description{
Creates QSAR datasets.
}
