% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_bioactivities.R
\name{downloadBioactivities}
\alias{downloadBioactivities}
\title{Downloads bioactivities from database}
\usage{
downloadBioactivities(.connection, .targets = NULL, .standard_type = NULL,
  .nmols_min = NULL, .nmols_max = NULL)
}
\arguments{
\item{.connection}{a database connection created using \code{dplyr::src_*}.}

\item{.targets}{a \code{Targets} object. Default: NULL.}

\item{.standard_type}{standard type identifier (character). Default: NULL.}

\item{.nmols_min}{min number of molecules per target (numeric). Default: NULL.}

\item{.nmols_max}{max number of molecules per target (numeric). Default: NULL.}
}
\value{
A \code{Bioactivities} object.
}
\description{
Downloads bioactivities from database.
}
