% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_fingerprints.R
\name{downloadFingerprints}
\alias{downloadFingerprints}
\title{Download fingerprints from the database}
\usage{
downloadFingerprints(.connection, .fingerprint_id, .molecule_id = NULL)
}
\arguments{
\item{.connection}{a database connection created using \code{dplyr::src_*}.}

\item{.fingerprint_id}{fingerprint identifier (character)}

\item{.molecule_id}{molecule identifiers (character vector). Default: NULL, all molecules.}
}
\value{
A \code{Fingerprint} object.
}
\description{
Download fingerprints from the database.
}
