% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_molrepresent.R
\name{downloadMolRepresentation}
\alias{downloadMolRepresentation}
\title{Downloads a molecular representation from the database}
\usage{
downloadMolRepresentation(.connection, .representation_id = NULL)
}
\arguments{
\item{.connection}{a database connection created using \code{dplyr::src_*}.}

\item{.representation_id}{representation identifier (numerical value). Default: NULL}
}
\value{
a \code{MolRepresentation} object.
}
\description{
Downloads a molecular representation from the database.
}
