% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_targets.R
\name{downloadTargets}
\alias{downloadTargets}
\title{Downloads targets from database.}
\usage{
downloadTargets(.connection)
}
\arguments{
\item{.connection}{a database connection created using \code{dplyr::src_*}.}
}
\value{
A \code{Targets} object.
}
\description{
Downloads targets form database.
}
