% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_fingerprints.R
\name{Fingerprint}
\alias{Fingerprint}
\title{Fingerprint class}
\usage{
Fingerprint(.fingerprint_id = NULL, .data = NULL, .nbits = NULL,
  .standard_base = NULL, .current_base = NULL)
}
\arguments{
\item{.fingerprint_id}{fingerprint identifier (character). Default: NULL}

\item{.data}{fingerprint representation of molecules (data.frame). Default: NULL}

\item{.nbits}{number of bits (numeric). Default: NULL}

\item{.standard_base}{standard fingerprint numeric base: 2,...,32 (numeric). Default: NULL}

\item{.current_base}{current numeric base: 2,...,32 (numeric). Default: NULL}
}
\value{
A \code{Fingerprint} object.
}
\description{
Creates a \code{Fingerprint} object.
}
