% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descriptors-generation.R
\name{generate_descriptors_from_sdf}
\alias{generate_descriptors_from_sdf}
\title{Generate Molecular Descriptors from SDF File}
\usage{
generate_descriptors_from_sdf(sdf_file)
}
\arguments{
\item{sdf_file}{Path to the SDF file.}
}
\value{
A matrix containing molecular descriptors for each molecule in the SDF file.
}
\description{
This function reads an SDF (Structure Data File) containing molecular structures and calculates molecular descriptors for each molecule.
}
