% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_molrepresent.R
\name{MolRepresentation}
\alias{MolRepresentation}
\title{MolRepresentation class}
\usage{
MolRepresentation(.representation_id = NULL, .fingerprint_id = NULL,
  .molproperty_id = NULL)
}
\arguments{
\item{.representation_id}{representation identifier (numerical value). Default: NULL}

\item{.fingerprint_id}{fingerprint identifiers (character vector). Default: NULL}

\item{.molproperty_id}{molecular property identifiers (character vector). Default: NULL}
}
\value{
A \code{MolRepresentation} object.
}
\description{
Creates a new \code{MolRepresentation} object.
}
