% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable-selection.R
\name{perform_variable_selection}
\alias{perform_variable_selection}
\title{Perform variable selection using regression subsets}
\usage{
perform_variable_selection(file_path, outcome_col, des_sel_meth = "exhaustive")
}
\arguments{
\item{file_path}{The file path of the dataset.}

\item{outcome_col}{The name of the outcome column.}

\item{des_sel_meth}{The method for variable selection (default is "exhaustive").}
}
\value{
A data frame containing the selected variables and the outcome.
}
\description{
This function performs variable selection using regression subsets method.
}
