% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_datasets.R
\name{QSARDatasets}
\alias{QSARDatasets}
\title{QSARDatasets class}
\usage{
QSARDatasets(.targets = NULL, .molrepresentation = NULL, .response = NULL,
  .datasets = NULL, .dataset_id = NULL)
}
\arguments{
\item{.targets}{a \code{Targets} object. Default: NULL.}

\item{.molrepresentation}{a \code{MolRepresentation} object. Default: NULL.}

\item{.response}{name of the response variable (character). Default: NULL.}

\item{.datasets}{named list of datasets (named list of data.frames). Default: NULL. Names should correspond to target ids.}

\item{.dataset_id}{dataset identifiers (numerical vector). Default: NULL.}
}
\value{
A \code{QSARDatasets} object.
}
\description{
Creates a \code{QSARDatasets} object.
}
