% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residual-plot.R
\name{residual_plots}
\alias{residual_plots}
\title{Function to create residual plots with model type labels}
\usage{
residual_plots(model_results)
}
\arguments{
\item{model_results}{A list containing model results}
}
\value{
A list of ggplot objects representing residual plots
}
\description{
Function to create residual plots with model type labels
}
