% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_targets.R
\name{Targets}
\alias{Targets}
\title{Targets class}
\usage{
Targets(.target_id = NULL)
}
\arguments{
\item{.target_id}{target identifiers (character vector). Default: NULL}
}
\value{
A \code{Targets} object.
}
\description{
Creates a new \code{Targets} object
}
