% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_uploaddb.R
\name{uploadDB.QSARDatasets}
\alias{uploadDB.QSARDatasets}
\title{Uploads QSAR datasets to the database}
\usage{
\method{uploadDB}{QSARDatasets}(.obj, .connection, ...)
}
\arguments{
\item{.obj}{a \code{QSARDatasets} object.}

\item{.connection}{a database connection created using \code{dplyr::src_*}.}

\item{...}{further arguments (currently not in use)}
}
\value{
a logical value that indicates whether the uploading was succesful.
}
\description{
Uploads a \code{QSARDatasets} object to the database.
}
