% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_uploaddb.R
\name{uploadDB}
\alias{uploadDB}
\title{Uploads an rQSAR object to the database}
\usage{
uploadDB(.obj, .connection)
}
\arguments{
\item{.obj}{an rQSAR object}

\item{.connection}{a database connection created using \code{dplyr::src_*}.}

\item{...}{further arguments (see specific functions for details)}
}
\value{
Usually a logical value indicating whether the uploading was succesful. See specific functions for more details.
}
\description{
Generic function to upload an rQSAR object to the database.
}
